[REFACTOR] [react] - Rename social-icons to icons.
