+++ packages/codistica-react/src/components/icons/index.js
